refactor: share one conflict backoff helper
